+++ src/steps/resource-manager/monitor/converters.ts
@@ -1,0 +1,33 @@
+import { AzureWebLinker } from '../../../azure';
+import {
+  createIntegrationEntity,
+  Entity,
+} from '@jupiterone/integration-sdk-core';
+import { MonitorEntities } from './constants';
+import { LogProfileResource } from '@azure/arm-monitor/esm/models';
+
+export function createMonitorLogProfileEntity(
+  webLinker: AzureWebLinker,
+  data: LogProfileResource,
+): Entity {
+  return createIntegrationEntity({
+    entityData: {
+      source: data,
+      assign: {
+        _key: data.id,
+        _type: MonitorEntities.MONITOR_LOG_PROFILE._type,
+        _class: MonitorEntities.MONITOR_LOG_PROFILE._class,
+        id: data.id,
+        webLink: webLinker.portalResourceUrl(data.id),
+        name: data.name,
+        displayName: data.name,
+        storageAccountId: data.storageAccountId,
+        serviceBusRuleId: data.serviceBusRuleId,
+        locations: data.locations,
+        categories: data.categories,
+        'retentionPolicy.enabled': data.retentionPolicy.enabled,
+        'retentionPolicy.days': data.retentionPolicy.days,
+      },
+    },
+  });
+}